Clean up gRPC storage config
