+++ plugin/storage/grpc/config_test.go
@@ -4,25 +4,10 @@
 package grpc
 
 import (
-	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/collector/component"
-	"google.golang.org/grpc"
 )
-
-func TestBuildRemoteNewClientError(t *testing.T) {
-	// this is a silly test to verify handling of error from grpc.NewClient, which cannot be induced via params.
-	c := &ConfigV2{}
-	newClientFn := func(_ ...grpc.DialOption) (conn *grpc.ClientConn, err error) {
-		return nil, errors.New("test error")
-	}
-	_, err := newRemoteStorage(c, component.TelemetrySettings{}, newClientFn)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "error creating remote storage client")
-}
 
 func TestDefaultConfigV2(t *testing.T) {
 	cfg := DefaultConfigV2()
